attachextract: refactor attachment text extractor from index.c
