+++ imap/index.c
@@ -86,6 +86,7 @@
 #include "seen.h"
 #include "statuscache.h"
 #include "strhash.h"
+#include "attachextract.h"
 #include "user.h"
 #include "util.h"
 #include "xstats.h"
@@ -100,8 +101,6 @@
 #include "imap/imap_err.h"
 
 EXPORTED unsigned client_capa;
-
-static struct extractor_ctx *index_text_extractor = NULL;
 
 /* Forward declarations */
 static void index_refresh_locked(struct index_state *state);
@@ -5292,13 +5291,6 @@ EXPORTED int index_search_evaluate(struct index_state *state,
     return match;
 }
 
-struct extractor_ctx {
-    struct protstream *clientin;
-    char *hostname;
-    char *path;
-    struct backend *be;
-};
-
 struct getsearchtext_rock
 {
     search_text_receiver_t *receiver;
@@ -5306,7 +5298,6 @@ struct getsearchtext_rock
     int charset_flags;
     const strarray_t *partids;
     int snippet_iteration; /* 0..no snippet, 1..first run, 2..second run */
-    struct extractor_ctx *ext;
     strarray_t striphtml; /* strip HTML from these plain text body part ids */
     uint8_t indexlevel;
     int flags;
@@ -5535,347 +5526,6 @@ done:
 
 #endif /* USE_HTTPD */
 
-
-#define IDLE_TIMEOUT (5 * 60)  /* 5 min */
-
-static int login(struct backend *s __attribute__((unused)),
-                 const char *userid __attribute__((unused)),
-                 sasl_callback_t *cb __attribute__((unused)),
-                 const char **status __attribute__((unused)),
-                 int noauth __attribute__((unused)))
-{
-    return 0;
-}
-
-static int ping(struct backend *s __attribute__((unused)),
-                const char *userid __attribute__((unused)))
-{
-    return 0;
-}
-
-static int logout(struct backend *s __attribute__((unused)))
-{
-    return 0;
-}
-
-static void extractor_disconnect(struct extractor_ctx *ext)
-{
-    if (!ext) return;
-
-    struct backend *be = ext->be;
-    syslog(LOG_DEBUG, "extractor_disconnect(%p)", be);
-
-    if (!be || (be->sock == -1)) {
-        /* already disconnected */
-        return;
-    }
-
-    /* need to logout of server */
-    backend_disconnect(be);
-
-    /* remove the timeout */
-    if (be->timeout) prot_removewaitevent(be->clientin, be->timeout);
-    be->timeout = NULL;
-    be->clientin = NULL;
-}
-
-static struct prot_waitevent *
-extractor_timeout(struct protstream *s __attribute__((unused)),
-                  struct prot_waitevent *ev __attribute__((unused)),
-                  void *rock)
-{
-    struct extractor_ctx *ext = rock;
-
-    syslog(LOG_DEBUG, "extractor_timeout(%p)", ext);
-
-    /* too long since we last used the extractor - disconnect */
-    extractor_disconnect(ext);
-
-    return NULL;
-}
-
-static struct protocol_t http =
-{ "http", "HTTP", TYPE_SPEC, { .spec = { &login, &ping, &logout } } };
-
-static int extractor_connect(struct extractor_ctx *ext)
-{
-    struct backend *be;
-    time_t now = time(NULL);
-
-    syslog(LOG_DEBUG, "extractor_connect()");
-
-    be = ext->be;
-    if (be && be->sock != -1) {
-        // extend the timeout
-        if (be->timeout) be->timeout->mark = now + IDLE_TIMEOUT;
-        return 0;
-    }
-
-    // clean up any existing connection
-    extractor_disconnect(ext);
-    be = ext->be = backend_connect(be, ext->hostname,
-                                   &http, NULL, NULL, NULL, -1);
-
-    if (!be) {
-        syslog(LOG_ERR, "extract_connect: failed to connect to %s",
-               ext->hostname);
-        return IMAP_IOERROR;
-    }
-
-    if (ext->clientin) {
-        /* add a default timeout */
-        be->clientin = ext->clientin;
-        be->timeout = prot_addwaitevent(ext->clientin,
-                                        now + IDLE_TIMEOUT,
-                                        extractor_timeout, ext);
-    }
-
-    return 0;
-}
-
-static int extract_attachment(const char *type, const char *subtype,
-                              const struct param *type_params,
-                              const struct buf *data, int encoding,
-                              const struct message_guid *content_guid,
-                              struct getsearchtext_rock *str)
-{
-    struct backend *be;
-    struct buf decbuf = BUF_INITIALIZER;
-    struct buf buf = BUF_INITIALIZER;
-    hdrcache_t hdrs = NULL;
-    struct body_t body = { 0, 0, 0, 0, 0, BUF_INITIALIZER };
-    const char *guidstr, *errstr = NULL;
-    size_t hostlen;
-    const char **hdr, *p;
-    int r = 0;
-
-    if (!index_text_extractor) {
-        /* This is a legitimate case for sieve and lmtpd (so we don't need
-         * to spam the logs! */
-        syslog(LOG_DEBUG, "%s: ignoring uninitialized extractor",
-                __func__);
-        return 0;
-    }
-
-    if (message_guid_isnull(content_guid)) {
-        syslog(LOG_DEBUG, "extract_attachment: ignoring null guid for %s/%s",
-               type ? type : "<null>", subtype ? subtype : "<null>");
-        return 0;
-    }
-
-    struct extractor_ctx *ext = str->ext = index_text_extractor;
-
-    r = extractor_connect(ext);
-    if (r) return r;
-    be = ext->be;
-
-    hostlen = strcspn(ext->hostname, "/");
-    guidstr = message_guid_encode(content_guid);
-
-    /* try to fetch previously extracted text */
-    unsigned statuscode = 0;
-    prot_printf(be->out,
-                "GET %s/%s %s\r\n"
-                "Host: %.*s\r\n"
-                "User-Agent: Cyrus/%s\r\n"
-                "Connection: Keep-Alive\r\n"
-                "Keep-Alive: timeout=%u\r\n"
-                "Accept: text/plain\r\n"
-                "X-Truncate-Length: " SIZE_T_FMT "\r\n"
-                "\r\n",
-                ext->path, guidstr, HTTP_VERSION,
-                (int) hostlen, be->hostname, CYRUS_VERSION,
-                IDLE_TIMEOUT, config_search_maxsize);
-    prot_flush(be->out);
-
-    /* Read GET response */
-    do {
-        r = http_read_response(be, METH_GET,
-                               &statuscode, &hdrs, &body, &errstr);
-        if (r) {
-            syslog(LOG_ERR,
-                   "extract_attachment: failed to read response for GET %s/%s",
-                   ext->path, guidstr);
-            statuscode = 599;
-        }
-    } while (statuscode < 200);
-
-    syslog(LOG_DEBUG, "extract_attachment: GET %s/%s: got status %u",
-           ext->path, guidstr, statuscode);
-
-    if (statuscode == 200) goto gotdata;
-
-    // otherwise we're going to try three times to PUT this request to the server!
-
-    /* Decode data */
-    if (encoding) {
-        if (charset_decode(&decbuf, buf_base(data), buf_len(data), encoding)) {
-            syslog(LOG_ERR, "extract_attachment: failed to decode data");
-            r = IMAP_IOERROR;
-            goto done;
-        }
-        data = &decbuf;
-    }
-
-    /* Build list of Content-Type parameters */
-    const struct param *param;
-    for (param = type_params; param && param->attribute; param = param->next) {
-        /* Ignore all but select parameters */
-        if (strcmp(param->attribute, "charset")) {
-            continue;
-        }
-        buf_putc(&buf, ';');
-        buf_appendcstr(&buf, param->attribute);
-        if (param->value) {
-            buf_putc(&buf, '=');
-            buf_appendcstr(&buf, param->value);
-        }
-    }
-
-    int retry;
-    for (retry = 0; retry < 3; retry++) {
-        if (retry) {
-            // second and third time around, sleep and reconnect
-            sleep(retry);
-            extractor_disconnect(ext);
-            r = extractor_connect(ext);
-            if (r) continue;
-            be = ext->be;
-        }
-
-        /* Send attachment to service for text extraction */
-        prot_printf(be->out,
-                    "PUT %s/%s %s\r\n"
-                    "Host: %.*s\r\n"
-                    "User-Agent: Cyrus/%s\r\n"
-                    "Connection: Keep-Alive\r\n"
-                    "Keep-Alive: timeout=%u\r\n"
-                    "Accept: text/plain\r\n"
-                    "Content-Type: %s/%s%s\r\n"
-                    "Content-Length: " SIZE_T_FMT "\r\n"
-                    "X-Truncate-Length: " SIZE_T_FMT "\r\n"
-                    "\r\n",
-                    ext->path, guidstr, HTTP_VERSION,
-                    (int) hostlen, be->hostname, CYRUS_VERSION, IDLE_TIMEOUT,
-                    type, subtype, buf_cstring(&buf), buf_len(data),
-                    config_search_maxsize);
-        prot_putbuf(be->out, data);
-        prot_flush(be->out);
-
-        /* Read PUT response */
-        body.flags = 0;
-        do {
-            r = http_read_response(be, METH_PUT,
-                                   &statuscode, &hdrs, &body, &errstr);
-            if (r) {
-                syslog(LOG_ERR,
-                       "extract_attachment: failed to read response for PUT %s/%s",
-                       ext->path, guidstr);
-                statuscode = 599;
-            }
-        } while (statuscode < 200);
-
-        syslog(LOG_DEBUG, "extract_attachment: PUT %s/%s: got status %u",
-               ext->path, guidstr, statuscode);
-
-        if (statuscode == 200 || statuscode == 201) {
-            // we got a result, yay
-            goto gotdata;
-        }
-
-        if (statuscode >= 400 && statuscode <= 499) {
-            /* indexer can't extract this for some reason, never try again */
-            goto done;
-        }
-
-        /* any other status code is an error */
-        syslog(LOG_ERR, "extract GOT STATUSCODE %d with timeout %d: %s", statuscode, IDLE_TIMEOUT, errstr);
-    }
-
-    // dropped out of the loop?  Then we failed!
-    r = IMAP_IOERROR;
-    goto done;
-
-gotdata:
-    /* Abide by server's timeout, if any */
-    if ((hdr = spool_getheader(hdrs, "Keep-Alive")) &&
-        (p = strstr(hdr[0], "timeout="))) {
-        int timeout = atoi(p+8);
-        if (be->timeout) be->timeout->mark = time(NULL) + timeout;
-    }
-    /* Append extracted text */
-    if (buf_len(&body.payload)) {
-        str->receiver->begin_part(str->receiver, SEARCH_PART_ATTACHMENTBODY);
-        str->receiver->append_text(str->receiver, &body.payload);
-        str->receiver->end_part(str->receiver, SEARCH_PART_ATTACHMENTBODY);
-    }
-
-done:
-    spool_free_hdrcache(hdrs);
-    buf_free(&body.payload);
-    buf_free(&buf);
-    buf_free(&decbuf);
-    return r;
-}
-
-EXPORTED void index_text_extractor_init(struct protstream *clientin)
-{
-    const char *exturl =
-         config_getstring(IMAPOPT_SEARCH_ATTACHMENT_EXTRACTOR_URL);
-    if (!exturl) return;
-
-    syslog(LOG_DEBUG, "extractor_init(%p)", clientin);
-
-    char scheme[6], server[100], path[256], *p;
-    unsigned https, port;
-
-    /* Parse URL (cheesy parser without having to use libxml2) */
-    int n = sscanf(exturl, "%5[^:]://%99[^/]%255[^\n]",
-                   scheme, server, path);
-    if (n != 3 ||
-        strncmp(lcase(scheme), "http", 4) || (scheme[4] && scheme[4] != 's')) {
-        syslog(LOG_ERR,
-               "extract_attachment: unexpected non-HTTP URL %s", exturl);
-        return;
-    }
-
-    /* Normalize URL parts */
-    https = (scheme[4] == 's');
-    if (*(p = path + strlen(path) - 1) == '/') *p = '\0';
-    if ((p = strrchr(server, ':'))) {
-        *p++ = '\0';
-        port = atoi(p);
-    }
-    else port = https ? 443 : 80;
-
-    /* Build servername, port, and options */
-    struct buf buf = BUF_INITIALIZER;
-    buf_printf(&buf, "%s:%u%s/noauth", server, port, https ? "/tls" : "");
-
-    index_text_extractor = xzmalloc(sizeof(struct extractor_ctx));
-    index_text_extractor->clientin = clientin;
-    index_text_extractor->path = xstrdup(path);
-    index_text_extractor->hostname = buf_release(&buf);
-}
-
-EXPORTED void index_text_extractor_destroy(void)
-{
-    struct extractor_ctx *ext = index_text_extractor;
-
-    syslog(LOG_DEBUG, "extractor_destroy(%p)", ext);
-
-    if (!ext) return;
-
-    extractor_disconnect(ext);
-    free(ext->be);
-    free(ext->hostname);
-    free(ext->path);
-    free(ext);
-
-    index_text_extractor = NULL;
-}
-
-
 static int getsearchtext_cb(int isbody, charset_t charset, int encoding,
                             const char *type, const char *subtype,
                             const struct param *type_params __attribute__((unused)),
@@ -5992,23 +5642,13 @@ static int getsearchtext_cb(int isbody, charset_t charset, int encoding,
             str->receiver->end_part(str->receiver, SEARCH_PART_BODY);
         }
     }
-    else if (isbody && (!strcmp(type, "APPLICATION") || !strcmp(type, "TEXT"))) {
-
 #ifdef USE_HTTPD
+    else if (isbody && (!strcmp(type, "APPLICATION") && !strcmp(subtype, "ICS"))) {
         // application/ics is an alias for text/icalendar
-        if (!strcmp(subtype, "ICS")) {
-            extract_icalbuf(data, charset, encoding, str);
-            goto done;
-        }
+        extract_icalbuf(data, charset, encoding, str);
+    }
 #endif /* USE_HTTPD */
-
-        // these are encrypted fields which aren't worth indexing
-        if (!strcmp(subtype, "PKCS7-MIME")) goto done;
-        if (!strcmp(subtype, "PKCS7-ENCRYPTED")) goto done;
-        if (!strcmp(subtype, "PKCS7-SIGNATURE")) goto done;
-        if (!strcmp(subtype, "PGP-SIGNATURE")) goto done;
-        if (!strcmp(subtype, "PGP-KEYS")) goto done;
-        if (!strcmp(subtype, "PGP-ENCRYPTED")) goto done;
+    else if (isbody && attachextract_supports_type(type, subtype)) {
 
         /* Ignore attachments in first snippet generation pass */
         if (str->snippet_iteration == 1) goto done;
@@ -6022,8 +5662,8 @@ static int getsearchtext_cb(int isbody, charset_t charset, int encoding,
             goto done;
         }
 
-        r = extract_attachment(type, subtype, type_params, data, encoding,
-                               content_guid, str);
+        r = attachextract_extract_part(type, subtype, type_params, data, encoding,
+                content_guid, str->receiver, SEARCH_PART_ATTACHMENTBODY);
         if (r) {
             syslog(LOG_ERR, "IOERROR index: can't extract attachment %s (%s/%s): %s",
                     message_guid_encode(content_guid),
