+++ imap/index.h
@@ -367,9 +367,7 @@ extern int index_reload_record(struct index_state *state,
                                uint32_t msgno,
                                struct index_record *record);
 
-extern void index_text_extractor_init(struct protstream *clientin);
-extern void index_text_extractor_destroy(void);
-
 extern int insert_into_mailbox_allowed(struct mailbox *mailbox);
+
 
 #endif /* INDEX_H */
